Fix error on Node 16 if ABI dir doesn't exist

+++ index.js
@@ -38,7 +38,9 @@ task(TASK_COMPILE, async function (args, hre, runSuper) {
   }
 
   if (config.clear) {
-    fs.rmdirSync(outputDirectory, { recursive: true });
+    if (fs.existsSync(outputDirectory)) {
+      fs.rmdirSync(outputDirectory, { recursive: true });
+    }
   }
 
   if (!fs.existsSync(outputDirectory)) {
